Put all classes into one package
