+++ src/main/java/duke/Deadline.java
@@ -1,20 +1,22 @@
+package duke;
+
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
 
 
-public class Event extends Task{
+public class Deadline extends Task{
     private LocalDate date;
     private LocalTime time;
 
-    public Event(String description, String deadline) {
-        super(description, "[E]");
+    public Deadline(String description, String deadline) {
+        super(description, "[D]");
         this.date = LocalDate.parse(deadline.substring(0,10));
         this.time = LocalTime.parse(deadline.substring(11), DateTimeFormatter.ofPattern("HHmm"));
     }
 
-    public Event(String description, String date, String time) {
-        super(description, "[E]");
+    public Deadline(String description, String date, String time) {
+        super(description, "[D]");
         this.date = LocalDate.parse(date, DateTimeFormatter.ofPattern("MMM d yyyy"));
         this.time = LocalTime.parse(time, DateTimeFormatter.ofPattern("h:ma"));
     }
@@ -22,9 +24,11 @@ public class Event extends Task{
     @Override
     public String getDescription() {
         return super.getDescription()
-                + " (at: "
+                + " (by: "
                 + this.date.format(DateTimeFormatter.ofPattern("MMM d yyyy"))
-                + this.time.format(DateTimeFormatter.ofPattern(" h:mma"))
+                + this.time.format(DateTimeFormatter.ofPattern(" h:ma"))
                 + ")";
     }
 }
+
+
